Adjust keygen and add advanced mode
